Add support for locating Mixin from the Mods directory.
I have explained my stance and reasoning on Mixin forks.
So I will not go into detail here.
However, this allows you to drop a fork in the mods folder and it will be loaded instead of the bundled one.
Also adds a synthetic 'mixin' mod that users can specify as a dependency in mods.toml